Deletes base_helper decorator, creates new base_helper, base_controller_decorator and corrects base_helper_spec.rb
