--- app/helpers/spree/base_helper_decorator.rb
@@ -1,21 +1,0 @@
-module Spree::BaseHelperDecorator
-  def display_volume_price(variant, quantity = 1, user = nil)
-    Spree::Money.new(
-      variant.volume_price(quantity, user),
-      currency: Spree::Config[:currency]
-    ).to_html
-  end
-
-  def display_volume_price_earning_percent(variant, quantity = 1, user = nil)
-    variant.volume_price_earning_percent(quantity, user).round.to_s
-  end
-
-  def display_volume_price_earning_amount(variant, quantity = 1, user = nil)
-    Spree::Money.new(
-      variant.volume_price_earning_amount(quantity, user),
-      currency: Spree::Config[:currency]
-    ).to_html
-  end
-end
-
-Spree::BaseHelper.prepend Spree::BaseHelperDecorator
